+++ widgets/rb-rating-helper.h
@@ -37,8 +37,8 @@ typedef struct _RBRatingPixbufs RBRatingPixbufs;
 
 gboolean rb_rating_render_stars (GtkWidget *widget, GdkWindow *window,
 				 RBRatingPixbufs *pixbufs,
-				 gulong x, gulong y,
-				 gulong x_offset, gulong y_offset,
+				 int x, int y,
+				 int x_offset, int y_offset,
 				 gdouble rating, gboolean selected);
 
 double   rb_rating_get_rating_from_widget (GtkWidget *widget,
